requirements.txt
werkzeug==2.3.6
python-dotenv
Flask[async]
Flask-SQLAlchemy
mysql-connector-python
Flask-Login
Flask-Migrate
pydantic
openai
pytest
bcrypt
Flask-WTF
wtforms-components
Flask-Caching